src/requirements.txt
Flask~=3.1.0
google-cloud-run>=0.10.0,<1.0.0
protobuf>=4.25.0,<5.0.0
gunicorn>=20.1.0,<21.0.0
